admin/requirements.txt
Flask
flask-pymongo
pyjwt
Flask-Cors
python-dotenv
virtualenv
